Add `mmrStoreType` to template
